feat: got safetensors saving working correctly for the clip model, in the save_pretrained method. Also got the saving of the config.json working correctly
